storage: move `Engine.ExportMVCCToSst` to `MVCCExportToSST`
This patch moves the interface method `Engine.ExportMVCCToSst()` to a
function `MVCCExportToSST()`, since there was no reason for this to be
implemented on the engine interface.

This is mostly a mechanical reorganization, except minor tweaks to
function parameters and return values, as well as comments.

Release note: None
